Major overhaul of the design sequence generation to account for missing loops
The addition of AAreference to augment the PDB.PDB class with insertion of residues gives the ability to generate SEQRES matching .pdb files. These can be exported for loop modelling of missing occupancy atoms and further manipulated for the sequence selection process to include correctly indexed loops and correctly identified protein ORF's.