huge rework of the internal atoms, residues, coords machinery
Due to issues copying structures with a dereferencing deepcopy or a useless shallow copy, implemented methods to shallow copy Structures for the purpose of returning the same Structure with new Coords. This is opposed to initializing a new object. In the process, all constituent groups are accessed by indices rather than type checking. This comes with Atom's now contained in a Atoms class and were stripped of their Coords (this may not be permanent due to write issues). Residue's are now in a Residues Class