remove generate fragments attempt inside of metrics
This was causing an issue where the first error missing fragments in a pose that was loaded was causing the attempt of generate_fragments which only added to the pose. This results in backbone clash errors from 3 entities in pose. Surprisingly, there should be 4 entities in the pose, but the first entity is not appear twice for some strange reason.