writing code to explore sub-symmetries in a pose
The modifications will consolidate the mechanism of finding the ASU and the intra-oligomeric contacts in a Nanohedra SCM with other input poses that only contain an ASU. This needs significant testing to ensure the matrix multiplications work as intended. Subsequent relocation of class properties such as number of atoms and coords to Model or SymmetricModel as outlined in comments. This should have no functionality on existing code due to inheritance