large overhaul of metrics
I switched around the order of a number of the metrics due to the processing of some metrics multiple times. Particularly, residue classifications were out of place and done in duplicate with the trajectory mean and with designs_by_protocol. Many of the final pose metrics are now extracted from the full scores/trajectory_df including expansion to include metrics besides just significance metrics in the protocol metrics columns. Divergence is done more coherently and percent of residues for the nanohedra/pose metrics are fixed so that they are not greater than 1. The double counting of fragments was removed by overwriting of fragments generated using symdesign generate fragments from fragments found during Nanohedra