addition of scouting to interface_design
along with scout.xml, the scouting procedure will allow metrics to be calculated from a single small FastDesign with core and rim pulsing. The result should show the interface design potential with out the costly high resolution sampling of the full design profile.
Additional improvements are to dynamically indicate that the input pdb (for instance the ASU) is refined already due to acquisition from the refined Database source. More work needs to be done to ensure that none Database structures input have the proper refinement or the input structures get separated in the composite oligomers for the refinement separately, then recreation of the ASU or another refined structure