process interface metrics on an entity basis
This obviates the need to explicitly describe the residues belonging to an interface1 or interface2. Since we only care about the unbound versus the bound state, the summation of all unbound measurements provides the comparison to the bound state. Getting these measurements is now flexible to any rosetta metrics output (indexed by the measurement number) instead of relying on a perfect interface split. This is compatible with both an interface split along the standard interface definition of "down the middle" and a per entity split where each entity is measured in isolation