handling msa data in class, refactoring
The use of a MultipleSequenceAlignment object instead of a custom dictionary object will facilitate the access and maintenance of this data type across module. Dealing with hydrophobic collapse and co-evolution information are two examples of improved processing and object permanence. Added, handling of SequenceProfile generation from hhblits using the reference sequence followed by adjustment of the profile to the atom sequences