Code draft with 3 big changes, needs debugging
First, initial ghost fragment euler lookup must validate for both ghosts of 1 to surface of 2, and surface of 1 to ghosts of 2. The overlap of both these lookups will provide a way to sidestep most of the ASU clashing checks later because only ghosts that are not clashing are used for search and the overlap must have both the forward and reverse ghost/surface pairs, indicating non-clashing overlap.
Second, all transforms and clashing is vectorized during Nanohedra docking which removes the intermitenent function with performs pose clash and scoring testing from the pose possibility production steps. Coupling all pose production steps with one another allows the entire group to become utilized in relation to the others relating clashing information and similarity of transforms to be clustered.
Third, the above vectorization enables all subsequent lookup operations to be vectorized and thus provides the possibility of a differentiable function on the docking landscape. This augmentation separates each of the individual operations on the asymmetric unit (pdb transformation and clash testing, guide coord transformation) and combines them into an entire matrix multiplication where in the cords and transforms are all stacked and the Transformations happen simultaneously versus nested in a for loop. Crucially, this prepares the code for the possible utilization of a single Nanohedra "energy function", where in, the Nanohedra score can be used to relate the transformation parameters to the desired docking outcome. Thus, producing a differentiable function to search for the most optimal docked configurations given a map of possible configurations