refactor set_atoms and set_residues to assign_atoms and assign_residues
These properly handle the initialization of Atoms and Residues from collections of Atom and Residue instances as well as residues derived from Chains and Entities. The dependent state is checked and reset_state and copy are performed if so. Analogous to cleaning the instances for reuse in new Structure objects. verify_coords ensures the correct lengths of Atoms and Coords containers