refactor Residue, Structure/derived class initialization. Major draft toward v0.5
Allow initialization by passing the parent Structure object which silently shares the required containers to dependent (child) StructureBase instances. All routes of initialization allow for proper setting of attributes across the scope of StructureBase inheritors. Current Structure initialization from Atoms and Coords is maintained allowing typical file parsing. create_residues, create_chains, and create_entities similarly utilize parent attributes to pass all necessary StructureBase containers. Finally, construction of Structure from_chains() and from_entities() classmethods work as previously, augmented by more thorough checking for the state of Atom and Residue instance use by existing Structures and performing a semi-deep copy and reset_state to create separate Structure instances with completely independent attributes so that manipulation of one doesn't affect the other. Atoms had fallen under the radar due to little manipulation, however, with the introduction of _sasa attribute, their copying could be prone to residual data carryover.

When each Structure is initiated on its own without constructor arguments, the default is to return an empty Residue/Structure object which would be quite useless, but viable. These should pair with methods to add_atoms() and add_residues() which need to be drafted. Along this ethos, setting .atoms and .residues attributes should be performed in the same manner as is .coords in the new style ".coords = new_coords" properly storing the new data in the parent StructureBase container but utilizing the calling Structure to perform the correct replacement.

The Structures still rely on __key() methods for equality comparison which are not detailed enough to distinguish two Structures with the same residues and name apart. If they were moved through cartesian space or have state attributes set, such as a transformation or sasa set, they would still evaluate as equal. In this way, one Structure initialized from the other may not have the predicted behavior if users want to explicitly differentiate between semi-deep copies of the same Structure. This is a design ethos that needs to be established, but at face value appears to be out of line with PEP recommendations.