refactoring and annotation
import modification based on switching file parsing from PDB.PDB to Pose.Model or Structure.Structure. Ensure structural method performed on the correct level of Structure. For instance, the refinement and loop modeling should be performed on the target input, regardless of entity number. The sequence based methods such as hhblits and msa generation should be based on each Entity