updated coords setting to account for empty container sets
The setting of coords from process_model is required to use self.coords = new_coords to account for Structure dependent handling of Coords containers. Previously coords would only be set to the new passed coords if none existed. This is the wrong behavior as if coords are provided they should be utilized