refactor nanohedra operations
The flags for starting nanohedra_dock have been reassigned to the JobResources object. All scripts and programs using nanohedra_dock have been modified accordingly. Key in this update is that distribution of Nanohedra has stopped working. I need to reconnect this using the Nanohedra.py script which will distribute the jobs that are set up and processed to begin by symdesign main