improve database set up and mapping
self-referential relationship between designs to create design hierarchies. Reconfigure Pose to have separate Pose*Metrics that are unique to the combination of EntityData in a single spatial orientation. All of these have UniqueConstraints and new rigorous UniqueConstraint checking during load of ProteinMetadata and UniProtEntity to ensure no new entries are added and all old entries are attached to new work